Module 2/Chapter1/django-myproject/requirements/dev.txt
-r base.txt

django-debug-toolbar